requirements.txt
aiohttp==3.7.4.post0
app==0.0.0
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-Login==0.5.0
idna==3.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
multidict==5.1.0
passlib==1.7.4
python-decouple==3.4
redis==3.5.3
soupsieve==2.2.1
typing-extensions==3.7.4.3
Werkzeug==1.0.1
yarl==1.6.3
